Move queue manager to notifier-seeder package
