+++ packages/commons/src/index.ts
@@ -11,4 +11,3 @@ export * from "./repositories/db.js";
 export * from "./types/index.js";
 export * from "./auth/token/index.js";
 export * from "./risk-analysis/index.js";
-export * from "./queue-manager/index.js";
