+++ packages/notifier-seeder/aws.config.local
@@ -1,0 +1,4 @@
+[default]
+aws_access_key_id=test-aws-key
+aws_secret_access_key=test-aws-secret
+region=eu-central-1
