+++ packages/notifier-seeder/package.json
@@ -9,6 +9,8 @@
     ".": "./dist/index.js"
   },
   "scripts": {
+    "test": "vitest",
+    "test:it": "vitest integration",
     "lint": "eslint . --ext .ts,.tsx",
     "lint:autofix": "eslint . --ext .ts,.tsx --fix",
     "format:check": "prettier --check src",
@@ -31,10 +33,15 @@
     "@types/uuid": "^9.0.2"
   },
   "dependencies": {
+    "@aws-sdk/client-sqs": "^3.529.1",
+    "dotenv-flow": "^3.2.0",
     "kafka-iam-auth": "workspace:*",
     "pagopa-interop-commons": "workspace:*",
     "pagopa-interop-models": "workspace:^",
+    "testcontainers": "^10.2.2",
     "ts-pattern": "^5.0.6",
-    "uuid": "^9.0.0"
+    "uuid": "^9.0.0",
+    "vitest": "^0.33.0",
+    "zod": "^3.0.0"
   }
 }
