+++ packages/notifier-seeder/src/index.ts
@@ -5,7 +5,6 @@ import {
   CatalogTopicConfig,
   catalogTopicConfig,
   getContext,
-  initQueueManager,
   kafkaConsumerConfig,
   logger,
   loggerConfig,
@@ -16,6 +15,7 @@ import { v4 as uuidv4 } from "uuid";
 import { kafkaMessageProcessError } from "pagopa-interop-models";
 import { convertToEserviceV1 } from "./models/catalogEventV2Converter.js";
 import { buildCatalogMessage } from "./models/catalogEventMessage.js";
+import { initQueueManager } from "./queue-manager/queueManager.js";
 
 const config = kafkaConsumerConfig();
 const topicsConfig = catalogTopicConfig();
