+++ packages/notifier-seeder/src/queue-manager/queueManagerErrors.ts
